+++ R/palette.R
@@ -14,13 +14,14 @@
 #' @param direction Either `1` or `-1`. If `-1` the palette will be reversed
 #' @param palette The name of the palette to sample from. See 
 #' [scico_palette_names()] for a list of possible names
+#' @param categorical Boolean. Should the categorical palettes be returned
 #' 
 #' @return A character vector of length `n` with hexencoded rgb(a) colour values
 #' 
 #' @references 
 #' <http://www.fabiocrameri.ch/colourmaps.php>
 #' 
-#' Crameri, Fabio. (2018, May 8). *Scientific colour maps (Version 3.0.1)*. Zenodo. \doi{10.5281/zenodo.1243909}
+#' Crameri, Fabio. (2021, September 12). *Scientific colour maps (Version 7.0.1)*. Zenodo. \doi{10.5281/zenodo.5501399}
 #' Crameri, Fabio. (2018). *Geodynamic diagnostics, scientific visualisation and StagLab 3.0*. Geosci. Model Dev. Discuss. \doi{10.5194/gmd-2017-328}
 #' 
 #' @export
@@ -37,7 +38,7 @@
 #' # Take a subset of a palette
 #' scico(15, begin = 0.3, end = 0.6, palette = 'berlin')
 #' 
-scico <- function(n, alpha = NULL, begin = 0, end = 1, direction = 1, palette = "bilbao") {
+scico <- function(n, alpha = NULL, begin = 0, end = 1, direction = 1, palette = "bilbao", categorical = FALSE) {
   if (begin < 0 | begin > 1 | end < 0 | end > 1) {
     stop("begin and end must be in [0,1]")
   }
@@ -52,7 +53,13 @@ scico <- function(n, alpha = NULL, begin = 0, end = 1, direction = 1, palette = 
     end <- tmp
   }
   
-  palette <- scico_palette_data(palette)
+  palette <- scico_palette_data(palette, categorical = categorical)
+  if (categorical) {
+    if (n > 100) {
+      stop("Cannot create categorical palettes with more than 100 colours")
+    }
+    return(palette[seq_len(n)])
+  }
   
   map_cols <- rgb(palette$r, palette$g, palette$b)
   fn_cols <- colorRamp(map_cols, space = "Lab", interpolate = "spline")
@@ -67,6 +74,8 @@ scico <- function(n, alpha = NULL, begin = 0, end = 1, direction = 1, palette = 
 #' 
 #' These functions gives access to the data underlying the palettes
 #' 
+#' @param categorical Boolean. Should the categorical palettes be returned
+#' 
 #' @keywords internal
 #' 
 #' @name scico_data
@@ -74,15 +83,27 @@ scico <- function(n, alpha = NULL, begin = 0, end = 1, direction = 1, palette = 
 #' 
 #' @export
 #' 
-scico_palette_names <- function() names(palettes)
+scico_palette_names <- function(categorical = FALSE) {
+  if (categorical) {
+    names(palettes$.categorical)
+  } else {
+    setdiff(names(palettes), ".categorical")
+  }
+}
 #' @rdname scico_data
 #' 
 #' @param palette The name of the palette to fetch data for
+#' @param categorical Boolean. Should the categorical palettes be returned
+#' 
 #' @export
 #' 
-scico_palette_data <- function(palette) {
-  palette <- match.arg(palette, names(palettes))
-  palettes[[palette]]
+scico_palette_data <- function(palette, categorical = FALSE) {
+  palette <- match.arg(palette, scico_palette_names(categorical))
+  if (categorical) {
+    palettes$.categorical[[palette]]
+  } else {
+    palettes[[palette]]
+  }
 }
 
 #' Show the different scico palettes
@@ -91,6 +112,8 @@ scico_palette_data <- function(palette) {
 #' available in the `scico` package
 #' 
 #' @param palettes One or more palette names to show
+#' @param categorical Boolean. Should the categorical palettes be returned
+#' @param n How many colours should be shown
 #' 
 #' 
 #' @importFrom grDevices n2mfrow
@@ -100,13 +123,14 @@ scico_palette_data <- function(palette) {
 #' @examples 
 #' 
 #' scico_palette_show()
+#' scico_palette_show(categorical = TRUE)
 #' 
-scico_palette_show <- function(palettes = scico_palette_names()) {
+scico_palette_show <- function(palettes = scico_palette_names(categorical), categorical = FALSE, n = if (categorical) 6 else 100) {
   dims <- n2mfrow(length(palettes))
   oldpar <- par(mfrow = dims, mai = par('mai')/5)
   on.exit(par(oldpar))
   
   for (i in palettes) {
-    image(matrix(1:100, ncol = 1), col = scico(100, palette = i), main = i, axes = FALSE)
+    image(matrix(seq_len(n), ncol = 1), col = scico(n, palette = i, categorical = categorical), main = i, axes = FALSE, )
   }
 }